Move getEntitySetSize to EntitySetManager, out of DataGraphManager remove hard deps on PostgresEntitySetSizesTask
